fix fetch status change

+++ src/app/api/mypageCreatePost/route.ts
@@ -12,7 +12,7 @@ export async function GET(req: Request) {
     const res = await sql`SELECT * FROM raid_posts WHERE user_id = ${userID}`
 
     return new Response(JSON.stringify({ postRows: res.rows }), {
-      status: 201,
+      status: 200,
     })
   } catch (error) {
     console.error(error)
@@ -23,12 +23,15 @@ export async function GET(req: Request) {
 export async function DELETE(req: Request) {
   const url = new URL(req.url)
   const post_id = url.searchParams.get('post_id')
+  if (!post_id) {
+    return new Response(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 404 })
+  }
 
   try {
     const res = await sql`DELETE FROM raid_posts WHERE post_id = ${post_id}`
 
     return new Response(JSON.stringify({ message: '성공' }), {
-      status: 201,
+      status: 200,
     })
   } catch (error) {
     console.error(error)
